+++ hutool-core/src/main/java/cn/hutool/core/bean/BeanInfoCache.java
@@ -1,8 +1,8 @@
 package cn.hutool.core.bean;
 
 import cn.hutool.core.lang.func.Func0;
-import cn.hutool.core.map.ReferenceConcurrentMap;
-import cn.hutool.core.map.WeakConcurrentMap;
+import cn.hutool.core.map.reference.ReferenceConcurrentMap;
+import cn.hutool.core.map.reference.WeakKeyValueConcurrentMap;
 
 import java.beans.PropertyDescriptor;
 import java.util.Map;
@@ -16,8 +16,8 @@ import java.util.Map;
 public enum BeanInfoCache {
 	INSTANCE;
 
-	private final WeakConcurrentMap<Class<?>, Map<String, PropertyDescriptor>> pdCache = new WeakConcurrentMap<>();
-	private final WeakConcurrentMap<Class<?>, Map<String, PropertyDescriptor>> ignoreCasePdCache = new WeakConcurrentMap<>();
+	private final WeakKeyValueConcurrentMap<Class<?>, Map<String, PropertyDescriptor>> pdCache = new WeakKeyValueConcurrentMap<>();
+	private final WeakKeyValueConcurrentMap<Class<?>, Map<String, PropertyDescriptor>> ignoreCasePdCache = new WeakKeyValueConcurrentMap<>();
 
 	/**
 	 * 获得属性名和{@link PropertyDescriptor}Map映射