+++ CHANGELOG.md
@@ -2,6 +2,22 @@
 # 🚀Changelog
 
 -------------------------------------------------------------------------------------------------------------
+# 5.8.41(2025-08-30)
+
+### 🐣新特性
+* 【core   】      增加`WeakKeyValueConcurrentMap`及其关联类，同时废弃`WeakConcurrentMap`并替换（issue#4039@Github）
+* 【core   】      `MapUtil`增加`removeByValue`和`removeIf`方法
+* 【core   】      `ObjectUtil`增加`apply`方法
+* 【core   】      `ReferenceUtil`增加`get`方法
+* 【db     】      `Condition`增加构造方法支持BETWEEN（issue#4041@Github）
+* 【core   】      `IoUtil.writeObjects`判空避免空指针（issue#4049@Github）
+
+### 🐞Bug修复
+* 【core   】      修复`ReflectUtil`中因class和Method关联导致的缓存无法回收问题（issue#4039@Github）
+* 【db     】      修复`Condition`的`Condition("discount_end_time", "!=", (String) null)`方法生成SQL时，生成SQL不符合预期要求的错误（pr#4042@Github）
+* 【core   】      修复`IoUtil`的`closeIfPosible`拼写错误，新建一个`closeIfPossible`方法，原方法标记deprecated（issue#4047@Github）
+
+-------------------------------------------------------------------------------------------------------------
 # 5.8.40(2025-08-26)
 
 ### 🐣新特性